+++ src/model/Plants/BlueTree.java
@@ -1,4 +1,4 @@
-package model;
+package model.Plants;
 
 public class BlueTree extends Plant{
