+++ src/model/Plants/GreenWheat.java
@@ -1,4 +1,4 @@
-package model;
+package model.Plants;
 
 public class GreenWheat extends Plant{
